# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API documentation",
        "description": "Документация",
        "version": "1.0.0"
    },
    "paths": {
        "/api/orders/{order_id}/worker": {
            "post": {
                "tags": [
                    "Order"
                ],
                "summary": "Привязка исполнителя к заказу",
                "description": "Возвращает пустой массив",
                "operationId": "pinWorker",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AttachEmployeeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                }
            }
        },
        "/api/orders/{order_id}": {
            "patch": {
                "tags": [
                    "Order"
                ],
                "summary": "Обновление статуса заказа",
                "description": "Обновляет статус закаказа",
                "operationId": "refresOrder",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/api/workers": {
            "get": {
                "tags": [
                    "Worker"
                ],
                "summary": "Получение списка испольнителей",
                "description": "Получение списка исполнителей",
                "operationId": "showWorkers",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "AttachEmployeeRequest": {
                "title": "Attact Worker to order",
                "description": "Привязка сотрудника к заказу",
                "required": [
                    "worker"
                ],
                "properties": {
                    "worker": {
                        "title": "worker",
                        "description": "Исполнитель",
                        "type": "string",
                        "example": "1"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Order",
            "description": "Order"
        },
        {
            "name": "Worker",
            "description": "Worker"
        }
    ]
}